feat: :package: add eslint, prettier, typescript
