+++ .prettierrc
@@ -1,0 +1,7 @@
+{
+  "printWidth": 80,
+  "tabWidth": 2,
+  "singleQuote": true,
+  "semi": true,
+  "trailingComma": "all"
+}
